# core/core-api/src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: soo-auth-service API
  description: 인증 서비스 API 가이드 (REST API Docs).
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost
...
paths:
...
  /api/v1/auth/admin/sms/logs:
    get:
      tags:
      - Admin SMS
      summary: 관리자 SMS 발송 이력 조회
      description: "발송 이력을 기간/페이지 조건으로 조회합니다.\n\n### 인증\n- `Authorization: Bearer\
        \ {accessToken}` 필요\n- 권한: `ROLE_ADMIN` 필요\n\n### 헤더\n- `Authorization`: Bearer\
        \ 관리자 액세스 토큰\n\n### Query 파라미터\n- `page`: 페이지 번호 (0부터 시작)\n- `size`: 페이지 크\
        기\n- `startDate`: 조회 시작 시각 (ISO-8601)\n- `endDate`: 조회 종료 시각 (ISO-8601)\n\n\
        ### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`: 응답\
        \ 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`:\
        \ 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n- `meta.request.method`: HTTP 메\
        서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n- `meta.paging`: 페이징 메타 정보\
        \ (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에\
        서만 포함)\n- `data`: 문자 발송 이력 페이지\n- `data.content`: 문자 발송 이력 목록\n- `data.content[].id`:\
        \ 이력 ID\n- `data.content[].to`: 수신 번호\n- `data.content[].from`: 발신 번호\n- `data.content[].text`:\
        \ 발송 메시지\n- `data.content[].ok`: 성공 여부\n- `data.content[].provider`: 발송 제공\
        자\n- `data.content[].code`: 실패 코드(없으면 null)\n- `data.content[].msg`: 실패 메시\
        지(없으면 null)\n- `data.content[].at`: 발송 시각\n- `data.totalElements`: 전체 건수\n\
        - `data.totalPages`: 전체 페이지 수\n- `data.size`: 페이지 크기\n- `data.number`: 현재\
        \ 페이지 번호\n- `data.numberOfElements`: 현재 페이지 요소 수\n- `data.first`: 첫 페이지 여부\
        \n- `data.last`: 마지막 페이지 여부\n- `data.empty`: 비어있는 페이지 여부\n- `data.pageable`:\
        \ null\n- `data.sort`: null\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n- **403**\
        \ 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시\
        는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-admin-sms-logs
      parameters:
      - name: page
        in: query
        description: 페이지 번호 (0부터 시작)
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: 페이지 크기
        required: false
        schema:
          type: string
      - name: startDate
        in: query
        description: 조회 시작 시각 (ISO-8601)
        required: false
        schema:
          type: string
      - name: endDate
        in: query
        description: 조회 종료 시각 (ISO-8601)
        required: false
        schema:
          type: string
      - name: Authorization
        in: header
        description: Bearer 관리자 액세스 토큰
        required: true
        schema:
          type: string
        example: Bearer token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-admin-sms-logs1758979252'
              examples:
                auth-admin-sms-logs:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:49.566815875\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/auth/admin/sms/logs\",\n      \"method\"\
                    \ : \"GET\",\n      \"query\" : \"page=0&size=20\"\n    }\n  },\n\
                    \  \"data\" : {\n    \"content\" : [ {\n      \"id\" : 1,\n  \
                    \    \"to\" : \"01012345678\",\n      \"from\" : \"029302266\"\
                    ,\n      \"text\" : \"[soo-auth] 인증번호 [123456] (유효 3분)\",\n  \
                    \    \"ok\" : true,\n      \"provider\" : \"SOLAPI\",\n      \"\
                    code\" : null,\n      \"msg\" : null,\n      \"at\" : \"2026-03-28T12:00:00\"\
                    \n    } ],\n    \"pageable\" : {\n      \"pageNumber\" : 0,\n\
                    \      \"pageSize\" : 20,\n      \"sort\" : {\n        \"empty\"\
                    \ : true,\n        \"sorted\" : false,\n        \"unsorted\" :\
                    \ true\n      },\n      \"offset\" : 0,\n      \"paged\" : true,\n\
                    \      \"unpaged\" : false\n    },\n    \"totalElements\" : 1,\n\
                    \    \"totalPages\" : 1,\n    \"last\" : true,\n    \"size\" :\
                    \ 20,\n    \"number\" : 0,\n    \"sort\" : {\n      \"empty\"\
                    \ : true,\n      \"sorted\" : false,\n      \"unsorted\" : true\n\
                    \    },\n    \"first\" : true,\n    \"numberOfElements\" : 1,\n\
                    \    \"empty\" : false\n  }\n}"
  /api/v1/auth/admin/sms/send:
    post:
      tags:
      - Admin SMS
      summary: 관리자 SMS 발송
      description: "관리자 권한으로 단문 메시지를 발송합니다.\n\n### 인증\n- `Authorization: Bearer {accessToken}`\
        \ 필요\n- 권한: `ROLE_ADMIN` 필요\n\n### 헤더\n- `Authorization`: Bearer 관리자 액세스 토\
        큰\n\n### 요청 바디 필드\n- `to`: 수신 번호\n- `text`: 발송 메시지\n\n### 응답 바디 필드\n- `result`:\
        \ 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`: 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`:\
        \ 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`: 처리 시간(ms)\n- `meta.request.path`:\
        \ 요청 경로\n- `meta.request.method`: HTTP 메서드\n- `meta.request.query`: 쿼리 문자열\
        \ (없으면 null)\n- `meta.paging`: 페이징 메타 정보 (있을 때만 반환)\n- `meta.locale`: 응답 로\
        케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에서만 포함)\n- `data`: 발송 결과\n- `data.id`:\
        \ 이력 ID\n- `data.to`: 수신 번호\n- `data.from`: 발신 번호\n- `data.text`: 발송 메시지\n\
        - `data.ok`: 성공 여부\n- `data.provider`: 발송 제공자\n- `data.at`: 발송 시각\n\n### 공\
        통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation\
        \ 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과\
        로 자동 포함됩니다."
      operationId: auth-admin-sms-send
      parameters:
      - name: Authorization
        in: header
        description: Bearer 관리자 액세스 토큰
        required: true
        schema:
          type: string
        example: Bearer token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-admin-sms-send-1110343501'
            examples:
              auth-admin-sms-send:
                value: "{\n  \"to\" : \"01011112222\",\n  \"text\" : \"관리자 공지 테스트\"\
                  \n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-admin-sms-send1615980123'
              examples:
                auth-admin-sms-send:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:50.41626951\",\n    \"request\" : {\n \
                    \     \"path\" : \"/api/v1/auth/admin/sms/send\",\n      \"method\"\
                    \ : \"POST\"\n    }\n  },\n  \"data\" : {\n    \"id\" : 2,\n \
                    \   \"to\" : \"01011112222\",\n    \"from\" : \"029302266\",\n\
                    \    \"text\" : \"관리자 공지 테스트\",\n    \"ok\" : true,\n    \"provider\"\
                    \ : \"SOLAPI\",\n    \"at\" : \"2026-03-28T13:00:00\"\n  }\n}"
  /api/v1/auth/admin/sms/stats:
    get:
      tags:
      - Admin SMS
      summary: 관리자 SMS 통계 조회
      description: "발송 성공/실패 통계를 조회합니다.\n\n### 인증\n- `Authorization: Bearer {accessToken}`\
        \ 필요\n- 권한: `ROLE_ADMIN` 필요\n\n### 헤더\n- `Authorization`: Bearer 관리자 액세스 토\
        큰\n\n### Query 파라미터\n- `startDate`: 조회 시작 시각 (ISO-8601)\n- `endDate`: 조회 종\
        료 시각 (ISO-8601)\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`:\
        \ 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`:\
        \ 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n- `meta.request.method`: HTTP 메\
        서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n- `meta.paging`: 페이징 메타 정보\
        \ (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에\
        서만 포함)\n- `data`: 문자 발송 통계\n- `data.total`: 전체 발송 수\n- `data.ok`: 성공 수\n-\
        \ `data.fail`: 실패 수\n- `data.rate`: 성공률(%)\n\n### 공통 에러\n- **401** 인증 실패/만\
        료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n\
        > 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-admin-sms-stats
      parameters:
      - name: startDate
        in: query
        description: 조회 시작 시각 (ISO-8601)
        required: false
        schema:
          type: string
      - name: endDate
        in: query
        description: 조회 종료 시각 (ISO-8601)
        required: false
        schema:
          type: string
      - name: Authorization
        in: header
        description: Bearer 관리자 액세스 토큰
        required: true
        schema:
          type: string
        example: Bearer token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-admin-sms-stats1143583724'
              examples:
                auth-admin-sms-stats:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:50.483984562\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/auth/admin/sms/stats\",\n      \"\
                    method\" : \"GET\"\n    }\n  },\n  \"data\" : {\n    \"total\"\
                    \ : 100,\n    \"ok\" : 95,\n    \"fail\" : 5,\n    \"rate\" :\
                    \ 95.0\n  }\n}"
...
components:
  schemas:
...
    api-v1-auth-admin-sms-stats1143583724:
      required:
      - data
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 처리 결과 (SUCCESS/ERROR)
        data:
          required:
          - fail
          - ok
          - rate
          - total
          type: object
          properties:
            fail:
              type: number
              description: 실패 수
            total:
              type: number
              description: 전체 발송 수
            rate:
              type: number
              description: 성공률(%)
            ok:
              type: number
              description: 성공 수
          description: 문자 발송 통계
        meta:
          required:
          - timestamp
          type: object
          properties:
            request:
              required:
              - method
              - path
              type: object
              properties:
                path:
                  type: string
                  description: 요청 경로
                method:
                  type: string
                  description: HTTP 메서드
                query:
                  type: string
                  description: 쿼리 문자열 (없으면 null)
                  nullable: true
            requestId:
              type: string
              description: 요청 추적 ID (있을 때만 반환)
              nullable: true
            paging:
              type: object
              description: 페이징 메타 정보 (있을 때만 반환)
              nullable: true
            locale:
              type: string
              description: 응답 로케일 (있을 때만 반환)
              nullable: true
            durationMs:
              type: number
              description: 처리 시간(ms)
              nullable: true
            timestamp:
              type: string
              description: 응답 생성 시각 (서버 로컬 시간)
        error:
          type: object
          description: 오류 정보 (에러 응답에서만 포함)
          nullable: true
...
    api-v1-auth-admin-sms-send1615980123:
      required:
      - data
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 처리 결과 (SUCCESS/ERROR)
        data:
          required:
          - at
          - from
          - id
          - ok
          - provider
          - text
          - to
          type: object
          properties:
            at:
              type: string
              description: 발송 시각
            provider:
              type: string
              description: 발송 제공자
            from:
              type: string
              description: 발신 번호
            text:
              type: string
              description: 발송 메시지
            to:
              type: string
              description: 수신 번호
            id:
              type: number
              description: 이력 ID
            ok:
              type: boolean
              description: 성공 여부
          description: 발송 결과
        meta:
          required:
          - timestamp
          type: object
          properties:
            request:
              required:
              - method
              - path
              type: object
              properties:
                path:
                  type: string
                  description: 요청 경로
                method:
                  type: string
                  description: HTTP 메서드
                query:
                  type: string
                  description: 쿼리 문자열 (없으면 null)
                  nullable: true
            requestId:
              type: string
              description: 요청 추적 ID (있을 때만 반환)
              nullable: true
            paging:
              type: object
              description: 페이징 메타 정보 (있을 때만 반환)
              nullable: true
            locale:
              type: string
              description: 응답 로케일 (있을 때만 반환)
              nullable: true
            durationMs:
              type: number
              description: 처리 시간(ms)
              nullable: true
            timestamp:
              type: string
              description: 응답 생성 시각 (서버 로컬 시간)
        error:
          type: object
          description: 오류 정보 (에러 응답에서만 포함)
          nullable: true
...
    api-v1-auth-admin-sms-send-1110343501:
      required:
      - text
      - to
      type: object
      properties:
        text:
          type: string
          description: 발송 메시지
        to:
          type: string
          description: 수신 번호
...
    api-v1-auth-admin-sms-logs1758979252:
      required:
      - data
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 처리 결과 (SUCCESS/ERROR)
        data:
          required:
          - content
          - empty
          - first
          - last
          - number
          - numberOfElements
          - size
          - totalElements
          - totalPages
          type: object
          properties:
            number:
              type: number
              description: 현재 페이지 번호
            last:
              type: boolean
              description: 마지막 페이지 여부
            numberOfElements:
              type: number
              description: 현재 페이지 요소 수
            size:
              type: number
              description: 페이지 크기
            totalPages:
              type: number
              description: 전체 페이지 수
            first:
              type: boolean
              description: 첫 페이지 여부
            content:
              type: array
              description: 문자 발송 이력 목록
              items:
                required:
                - at
                - from
                - id
                - ok
                - provider
                - text
                - to
                type: object
                properties:
                  msg:
                    type: object
                    description: 실패 메시지(없으면 null)
                    nullable: true
                  at:
                    type: string
                    description: 발송 시각
                  code:
                    type: object
                    description: 실패 코드(없으면 null)
                    nullable: true
                  provider:
                    type: string
                    description: 발송 제공자
                  from:
                    type: string
                    description: 발신 번호
                  text:
                    type: string
                    description: 발송 메시지
                  to:
                    type: string
                    description: 수신 번호
                  id:
                    type: number
                    description: 이력 ID
                  ok:
                    type: boolean
                    description: 성공 여부
            empty:
              type: boolean
              description: 비어있는 페이지 여부
            totalElements:
              type: number
              description: 전체 건수
          description: 문자 발송 이력 페이지
        meta:
          required:
          - timestamp
          type: object
          properties:
            request:
              required:
              - method
              - path
              type: object
              properties:
                path:
                  type: string
                  description: 요청 경로
                method:
                  type: string
                  description: HTTP 메서드
                query:
                  type: string
                  description: 쿼리 문자열 (없으면 null)
                  nullable: true
            requestId:
              type: string
              description: 요청 추적 ID (있을 때만 반환)
              nullable: true
            paging:
              type: object
              description: 페이징 메타 정보 (있을 때만 반환)
              nullable: true
            locale:
              type: string
              description: 응답 로케일 (있을 때만 반환)
              nullable: true
            durationMs:
              type: number
              description: 처리 시간(ms)
              nullable: true
            timestamp:
              type: string
              description: 응답 생성 시각 (서버 로컬 시간)
        error:
          type: object
          description: 오류 정보 (에러 응답에서만 포함)
          nullable: true
...